pyproject.toml
[project]
name = "ai-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "cohere>=5.21.1",
    "fastapi>=0.135.2",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
